simple cte's work serially now

+++ src/include/execution/ast/builtins.h
@@ -356,7 +356,7 @@ class Builtins {
   /**
    * @return The name of the function associated with the given builtin enumeration.
    */
-  static const char *GetFunctionName(Builtin builtin) { return builtin_function_names[static_cast<uint8_t>(builtin)]; }
+  static const char *GetFunctionName(Builtin builtin) { return builtin_function_names[static_cast<uint16_t>(builtin)]; }
 
  private:
   static const char *builtin_function_names[];
